Make changes to MainActivity startup logic to properly pull from Context
