+++ SeQR/app/build.gradle.kts
@@ -36,31 +36,20 @@ android {
 }
 
 dependencies {
-    implementation("androidx.slidingpanelayout:slidingpanelayout:1.2.0")
+    //implementation("androidx.slidingpanelayout:slidingpanelayout:1.2.0")
     implementation("com.google.firebase:firebase-storage:20.3.0")
-    val cameraxVersion = "1.2.2"
-    implementation("androidx.camera:camera-core:${cameraxVersion}")
-    implementation("androidx.camera:camera-camera2:${cameraxVersion}")
-    implementation("androidx.camera:camera-lifecycle:${cameraxVersion}")
-    implementation("androidx.camera:camera-video:${cameraxVersion}")
-    implementation("androidx.camera:camera-view:${cameraxVersion}")
-    implementation("androidx.camera:camera-extensions:${cameraxVersion}")
-    implementation("com.journeyapps:zxing-android-embedded:4.1.0")
     implementation("com.google.android.gms:play-services-code-scanner:16.1.0")
-    implementation("androidx.appcompat:appcompat:1.6.1")
-    implementation("com.google.android.material:material:1.11.0")
     implementation("androidx.constraintlayout:constraintlayout:2.1.4")
     implementation(platform("com.google.firebase:firebase-bom:32.7.1"))
     implementation("com.google.firebase:firebase-firestore")
     implementation("com.squareup.picasso:picasso:2.71828")
+    implementation("androidx.appcompat:appcompat:1.6.1")
+    implementation("com.google.android.material:material:1.11.0")
     testImplementation("junit:junit:4.13.2")
     androidTestImplementation("androidx.test.ext:junit:1.1.5")
     androidTestImplementation("androidx.test.espresso:espresso-core:3.5.1")
-
-
     //image upload stuff
-    implementation("com.squareup.picasso:picasso:2.5.2")
-
+    //implementation("com.squareup.picasso:picasso:2.5.2")
     implementation("com.journeyapps:zxing-android-embedded:4.1.0")
 
 }
