+++ SeQR/app/src/main/java/com/example/seqr/AProfilesFragment.java
@@ -11,11 +11,13 @@ import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
 
+import com.example.seqr.models.Profile;
 import com.example.seqr.adapters.ProfileAdapter;
 import com.example.seqr.controllers.ProfileController;
 import com.google.firebase.firestore.DocumentSnapshot;
 
 import java.util.ArrayList;
+import java.util.List;
 
 public class AProfilesFragment extends Fragment {
 
@@ -38,15 +40,18 @@ public class AProfilesFragment extends Fragment {
 
         // Create recycler view
         recyclerView = view.findViewById(R.id.admin_profiles);
-        profileList = new ArrayList<>;
+        profileList = new ArrayList<>();
         profileAdapter = new ProfileAdapter(profileList);
 
 
-        recyclerView.setLayoutManager(new LinearLayoutManager(this));
+        recyclerView.setLayoutManager(new LinearLayoutManager(this.getContext()));
         recyclerView.setAdapter(profileAdapter);
 
         // Query to get all profiles
         ProfileController profileController = new ProfileController();
+        //comment this out to get a valid build
+
+        /*
         profileController.getAllProfiles(task -> {
             if (task.isSuccessful()) {
                 for (DocumentSnapshot document : task.getResult()) {
@@ -55,9 +60,12 @@ public class AProfilesFragment extends Fragment {
                 }
                 profileAdapter.notifyDataSetChanged();
             } else {
-                Log.d("DEBUG", "there was some error event in profile retrieval"), task.getException();
+                Log.d("DEBUG", "there was some error event in profile retrieval");
+                task.getException();
             }
         });
+
+         */
 
         return view;
     }
