+++ SeQR/app/src/main/res/layout/activity_main.xml
@@ -1,14 +1,10 @@
 <?xml version="1.0" encoding="utf-8"?>
-
-
-
 <androidx.drawerlayout.widget.DrawerLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     android:id="@+id/my_drawer_layout"
     android:layout_width="match_parent"
     android:layout_height="match_parent">
-
 
     <RelativeLayout
     android:layout_width="match_parent"
